auto_executor/requirements.txt
pyautogui
pynput
Pillow
pytesseract
opencv-python
pyyaml
pytest
